fix(test): reduce startup-heavy hotspot retention (#52381)
Extract applyMSTeamsWebhookTimeouts + constants from monitor.ts into
standalone webhook-timeouts.ts module and update test import. Subset
of upstream dbd26e49f1 — fork takes only the msteams-local changes.

(cherry picked from commit dbd26e49f1336221c25cb0541ed56d0df74f8224)
